Updated routes to find device being used, BlankUser now has dynamic width and height, Nav shows gear heroicon instead of profile icon

+++ routes/web.php
@@ -4,6 +4,7 @@ use App\Http\Controllers\ProfileController;
 use Illuminate\Foundation\Application;
 use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
+use Jenssegers\Agent\Facades\Agent;
 
 /*
 |--------------------------------------------------------------------------
@@ -16,29 +17,32 @@ use Inertia\Inertia;
 |
 */
 
-Route::get('/', function () {
-    return Inertia::render('Auth/Login');
-});
-
-Route::get('/dashboard', function () {
-    return Inertia::render('Dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
-
-Route::middleware('auth')->group(function () {
-    Route::get('/{user:username}/profile', [ProfileController::class, 'edit'])->name('profile.edit');
-    Route::patch('/{user:username}/profile', [ProfileController::class, 'update'])->name('profile.update');
-    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
-
-    Route::prefix('contacts')->group(function () {
-        Route::get('/contacts-list', function () {
-            return Inertia::render('Contacts/Contacts');
-        })->name('contacts.list');
-
-        Route::get('/contact/edit', function () {
-            return Inertia::render('Contacts/ContactEdit');
-        })->name('contact.edit');
+if (!Agent::isMobile()) {
+    dd('Please change view go to mobile.');
+} else {
+    Route::get('/', function () {
+        return Inertia::render('Auth/Login');
     });
-});
 
+    Route::get('/dashboard', function () {
+        return Inertia::render('Dashboard');
+    })->middleware(['auth', 'verified'])->name('dashboard');
+
+    Route::middleware('auth')->group(function () {
+        Route::get('/{user:username}/profile', [ProfileController::class, 'edit'])->name('profile.edit');
+        Route::patch('/{user:username}/profile', [ProfileController::class, 'update'])->name('profile.update');
+        Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+
+        Route::prefix('contacts')->group(function () {
+            Route::get('/contacts-list', function () {
+                return Inertia::render('Contacts/Contacts');
+            })->name('contacts.list');
+
+            Route::get('/contact/edit', function () {
+                return Inertia::render('Contacts/ContactEdit');
+            })->name('contact.edit');
+        });
+    });
+}
 
 require __DIR__ . '/auth.php';
